Move Every to gigantism and put log fan-out in Loggable's companion
Relocate `Every` (case class + macro) from `frontier` to `gigantism`, a
bottom-of-stack module that already carries the scala3-compiler dependency
the `searchIgnoring`-based collection macro needs. `frontier` no longer
exports it; its existing users now depend on `gigantism` directly.

With `Every` reachable from `anticipation.log`, introduce a `Sink` trait
there and move the fan-out derivation into `Loggable`'s companion (generic
over the carrier, to avoid a cycle through fulminate's `Message`). Because a
companion-object given is in implicit scope, `Log.fine/info/warn/fail` now
resolve `event is Loggable` with no import at the use site; `eucalyptus.Logger`
becomes a `Sink` and drops its `frontier` dependency.
