+++ lib/anticipation/src/log/anticipation.Loggable.scala
@@ -34,13 +34,23 @@ package anticipation
 
 import language.experimental.into
 
+import gigantism.Every
 import prepositional.*
 
 object Loggable:
-  given transcribable: [input: Loggable, output: Transcribable to input] => output is Loggable =
+  // Derives the single `event is Loggable` that `Log.fine`/`info`/`warn`/`fail` summon: transcribe
+  // the event to a common `carrier` once, then fan out to EVERY in-scope `Sink` for that carrier
+  // (contravariance means a `Sink[Any, carrier]` is collected for any event). No sink in scope ⇒ an
+  // empty `Every` ⇒ silent; many ⇒ fan-out with per-sink routing. Living in `Loggable`'s companion
+  // keeps it in implicit scope, so no import is needed at the use site.
+  given fanOut: [event, carrier]
+  =>  ( transcribable: event is Transcribable to carrier, sinks: Every[Sink[event, carrier]] )
+  =>  event is Loggable =
+
     (level, timestamp, event) =>
-      if !output.skip(event)
-      then input.log(level, timestamp, output.record(event))
+      if !transcribable.skip(event) then
+        val message = transcribable.record(event)
+        sinks.values.foreach(_.submit(level, timestamp, message))
 
 trait Loggable extends Typeclass:
   loggable =>
