+++ lib/frontier/src/core/frontier.internal.scala
@@ -547,47 +547,3 @@ object internal:
 
       case a: Available =>
         emit(Missing(stenography.internal.name[target], List(a), Nil))
-
-  def every[value: Type]: Macro[Every[value]] =
-    import quotes.reflect.*
-
-    given context: core.Contexts.Context = quotes.absolve match
-      case quotes: runtime.impl.QuotesImpl => quotes.ctx
-
-    def underlying(tree: Term): Symbol = tree match
-      case Inlined(_, _, body) => underlying(body)
-      case Apply(fun, _)       => underlying(fun)
-      case TypeApply(fun, _)   => underlying(fun)
-      case Block(_, expr)      => underlying(expr)
-      case Typed(expr, _)      => underlying(expr)
-      case other               => other.symbol
-
-    def collect(ignored: List[Symbol], acc: List[Expr[value]]): List[Expr[value]] =
-      Implicits.searchIgnoring(TypeRepr.of[value])(ignored*).absolve match
-        case success: ImplicitSearchSuccess =>
-          val symbol = underlying(success.tree)
-
-          if symbol.isNoSymbol || ignored.contains(symbol) then acc.reverse
-          else collect(symbol :: ignored, success.tree.asExprOf[value] :: acc)
-
-        case failure: ImplicitSearchFailure =>
-          failure.asInstanceOf[ast.tpd.Tree].tpe match
-            case ambiguous: typer.Implicits.AmbiguousImplicits =>
-              val tree1 = ambiguous.alt1.tree.asInstanceOf[Term]
-              val tree2 = ambiguous.alt2.tree.asInstanceOf[Term]
-              val symbol1 = underlying(tree1)
-              val symbol2 = underlying(tree2)
-
-              val unusable =
-                symbol1.isNoSymbol || symbol2.isNoSymbol ||
-                  ignored.contains(symbol1) || ignored.contains(symbol2)
-
-              if unusable then acc.reverse
-              else collect
-                ( symbol1 :: symbol2 :: ignored,
-                  tree2.asExprOf[value] :: tree1.asExprOf[value] :: acc )
-
-            case _ =>
-              acc.reverse
-
-    '{Every[value](${Expr.ofList(collect(Nil, Nil))})}
