+++ lib/frontier/src/core/frontier_core.scala
@@ -32,7 +32,5 @@
                                                                                                   */
 package frontier
 
-inline def every[value]: Every[value] = ${internal.every[value]}
-
 package context:
   transparent inline given explainMissingContext: [any] => any = internal.explanation[any]
