+++ lib/frontier/src/core/soundness_frontier_core.scala
@@ -36,7 +36,6 @@ import frontier.*
 
 // unexported: Diagnostic (clashes with `harlequin.Diagnostic` in the umbrella; reach it via
 // `frontier.Diagnostic`)
-export frontier.Every
 
 package context:
   transparent inline given explainMissingContext: [any] => any = internal.explanation[any]
